# openapi.yaml
openapi: 3.1.0
info:
  title: Домовой — API приложения дома в MAX
  version: 1.0.0
  description: 'API мини-приложения и бота MAX «Домовой»: вход по QR-коду квитанции, обращения в УК, начисления,
    счётчики, лента и опросы дома, раздел председателя, кабинет УК, подбор дома.


    **Как войти для проверки.** Без квитанции и без MAX — ролью демо-дома: `POST /api/demo/roles/{key}/take` (key:
    chairman, owner12, member12, owner45, owner78, newcomer30) вернёт `token`; дальше заголовок `Authorization:
    Bearer <token>`. Кабинет УК — `POST /api/dispatcher/login` с логином демо-УК (пароль — на служебном слайде презентации),
    тоже `token`.


    **Ошибки** — JSON `{ "error": "код", "message": "текст для человека" }`.


    **Модель доступа.** Квитанция — заявка, а не пропуск: новая квартира приходит со статусом `pending`. Уровень
    0 (pending) — свои начисления, счётчики и обращения в УК сразу. Уровень 1 (active) — соседи, лента, опросы.
    Уровень 2 — дом целиком, только председателю. Обращение удалить нельзя.


    Кабинет оператора (`/api/admin/*`) — внутренний и здесь не описан.'
servers:
- url: https://domovoy.devcore.com.ru
  description: Бой
- url: http://localhost:3000
  description: Локальный запуск (docker compose)
...
security:
- bearerAuth: []
- cookieAuth: []
paths:
...
  /api/requests:
    get:
      tags:
      - Обращения
      summary: Мои обращения
      description: Активные и архив по квартире. Обращение нельзя удалить — маршрута удаления нет.
      operationId: get_api_requests
      parameters:
      - name: propertyId
        in: query
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  active:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        number:
                          type: string
                        kind:
                          type:
                          - string
                          - 'null'
                        category:
                          type:
                          - string
                          - 'null'
                        title:
                          type:
                          - string
                          - 'null'
                        description:
                          type:
                          - string
                          - 'null'
                        status:
                          type: string
                        statusLabel:
                          type:
                          - string
                          - 'null'
                        closed:
                          type:
                          - boolean
                          - 'null'
                        awaitingResident:
                          type:
                          - boolean
                          - 'null'
                        sla:
                          type:
                          - string
                          - 'null'
                        slaLabel:
                          type:
                          - string
                          - 'null'
                        hasDeadline:
                          type:
                          - boolean
                          - 'null'
                        assigneeName: {}
                        rejectReason: {}
                        createdAt:
                          type:
                          - string
                          - 'null'
                          format: date-time
                        closedAt:
                          type:
                          - string
                          - 'null'
                          format: date-time
                        masterSlotStart: {}
                        masterSlotEnd: {}
                      required:
                      - id
                      - number
                      - kind
                      - category
                      - title
                      - description
                      - status
                      - statusLabel
                      - closed
                      - awaitingResident
                      - sla
                      - slaLabel
                      - hasDeadline
                      - assigneeName
                      - rejectReason
                      - createdAt
                      - closedAt
                      - masterSlotStart
                      - masterSlotEnd
                  archive:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        number:
                          type: string
                        kind:
                          type:
                          - string
                          - 'null'
                        category:
                          type:
                          - string
                          - 'null'
                        title:
                          type:
                          - string
                          - 'null'
                        description:
                          type:
                          - string
                          - 'null'
                        status:
                          type: string
                        statusLabel:
                          type:
                          - string
                          - 'null'
                        closed:
                          type:
                          - boolean
                          - 'null'
                        awaitingResident:
                          type:
                          - boolean
                          - 'null'
                        sla:
                          type:
                          - string
                          - 'null'
                        slaLabel:
                          type:
                          - string
                          - 'null'
                        hasDeadline:
                          type:
                          - boolean
                          - 'null'
                        assigneeName: {}
                        rejectReason: {}
                        createdAt:
                          type:
                          - string
                          - 'null'
                          format: date-time
                        closedAt:
                          type:
                          - string
                          - 'null'
                          format: date-time
                        masterSlotStart: {}
                        masterSlotEnd: {}
                      required:
                      - id
                      - number
                      - kind
                      - category
                      - title
                      - description
                      - status
                      - statusLabel
                      - closed
                      - awaitingResident
                      - sla
                      - slaLabel
                      - hasDeadline
                      - assigneeName
                      - rejectReason
                      - createdAt
                      - closedAt
                      - masterSlotStart
                      - masterSlotEnd
                  archiveTotal:
                    type:
                    - number
                    - 'null'
                required:
                - active
                - archive
                - archiveTotal
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - Обращения
      summary: Подать обращение в УК
      description: 'Ядро продукта: работает сразу, до подтверждения жителя. Проверка на повтор, срок реакции по
        категории.'
      operationId: post_api_requests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                propertyId:
                  type:
                  - string
                  - 'null'
                category:
                  type:
                  - string
                  - 'null'
                description:
                  type:
                  - string
                  - 'null'
                draftId:
                  type:
                  - string
                  - 'null'
                kind:
                  type:
                  - string
                  - 'null'
                slotStart:
                  type:
                  - string
                  - 'null'
                  format: date-time
                slotEnd:
                  type:
                  - string
                  - 'null'
                  format: date-time
              required:
              - propertyId
              - description
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  id:
                    type: string
                  number:
                    type: string
                  slaHours:
                    type:
                    - number
                    - 'null'
                required:
                - status
                - id
                - number
                - slaHours
        '400':
          description: Неверные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Нет прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/requests/{id}:
    get:
      tags:
      - Обращения
      summary: Карточка обращения
      description: Статус, срок, переписка с УК, вложения.
      operationId: get_api_requests_id
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  number:
                    type: string
                  kind:
                    type:
                    - string
                    - 'null'
                  category:
                    type:
                    - string
                    - 'null'
                  title:
                    type:
                    - string
                    - 'null'
                  description:
                    type:
                    - string
                    - 'null'
                  status:
                    type: string
                  statusLabel:
                    type:
                    - string
                    - 'null'
                  closed:
                    type:
                    - boolean
                    - 'null'
                  awaitingResident:
                    type:
                    - boolean
                    - 'null'
                  sla:
                    type:
                    - string
                    - 'null'
                  slaLabel:
                    type:
                    - string
                    - 'null'
                  hasDeadline:
                    type:
                    - boolean
                    - 'null'
                  assigneeName:
                    type:
                    - string
                    - 'null'
                  rejectReason:
                    type:
                    - string
                    - 'null'
                  createdAt:
                    type:
                    - string
                    - 'null'
                    format: date-time
                  closedAt:
                    type:
                    - string
                    - 'null'
                    format: date-time
                  masterSlotStart:
                    type:
                    - string
                    - 'null'
                    format: date-time
                  masterSlotEnd:
                    type:
                    - string
                    - 'null'
                    format: date-time
                  addressee:
                    type: object
                    properties:
                      kind:
                        type:
                        - string
                        - 'null'
                      name:
                        type:
                        - string
                        - 'null'
                      phone: {}
                      hasCabinet:
                        type:
                        - boolean
                        - 'null'
                    required:
                    - kind
                    - name
                    - phone
                    - hasCabinet
                  events:
                    type: array
                    items:
                      type: object
                      properties:
                        text:
                          type:
                          - string
                          - 'null'
                        actor:
                          type:
                          - string
                          - 'null'
                        actorName:
                          type:
                          - string
                          - 'null'
                        type:
                          type:
                          - string
                          - 'null'
                        at:
                          type:
                          - string
                          - 'null'
                          format: date-time
                      required:
                      - text
                      - actor
                      - actorName
                      - type
                      - at
                  photos:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type:
                          - string
                          - 'null'
                        mime:
                          type:
                          - string
                          - 'null'
                        sizeBytes:
                          type:
                          - number
                          - 'null'
                        byDispatcher:
                          type:
                          - boolean
                          - 'null'
                        at:
                          type:
                          - string
                          - 'null'
                          format: date-time
                        url:
                          type:
                          - string
                          - 'null'
                      required:
                      - id
                      - name
                      - mime
                      - sizeBytes
                      - byDispatcher
                      - at
                      - url
                  rating:
                    type:
                    - object
                    - 'null'
                    properties:
                      stars:
                        type:
                        - number
                        - 'null'
                      comment:
                        type:
                        - string
                        - 'null'
                    required:
                    - stars
                    - comment
                required:
                - id
                - number
                - kind
                - category
                - title
                - description
                - status
                - statusLabel
                - closed
                - awaitingResident
                - sla
                - slaLabel
                - hasDeadline
                - assigneeName
                - rejectReason
                - createdAt
                - closedAt
                - masterSlotStart
                - masterSlotEnd
                - addressee
                - events
                - photos
                - rating
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/requests/{id}/comment:
    post:
      tags:
      - Обращения
      summary: Дописать в обращение
      description: Ответ на вопрос диспетчера или дополнение.
      operationId: post_api_requests_id_comment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type:
                  - string
                  - 'null'
              required:
              - text
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  reopened:
                    type:
                    - boolean
                    - 'null'
                required:
                - status
                - reopened
        '400':
          description: Неверные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Конфликт состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/requests/{id}/dispute:
    post:
      tags:
      - Обращения
      summary: Проблема не решена
      description: Вернуть выполненное обращение в работу — в течение 30 дней.
      operationId: post_api_requests_id_dispute
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    const: in_work
                required:
                - status
        '400':
          description: Неверные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Нет прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Конфликт состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  description: Что не так
...
  /api/requests/{id}/files/{fileId}:
    get:
      tags:
      - Обращения
      summary: Скачать вложение
      description: Только тому, кому видно само обращение.
      operationId: get_api_requests_id_files_fileId
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: fileId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Успех
          content:
            image/png:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Токен сессии из /api/auth/qr, /api/auth/max, /api/demo/roles/{key}/take или /api/dispatcher/login
    cookieAuth:
      type: apiKey
      in: cookie
      name: zd_session
      description: Та же сессия в cookie — её ставит сервер при входе
...
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
          description: Машинный код ошибки
        message:
          type: string
          description: Понятный человеку текст
  responses:
    Unauthorized:
      description: Нужен вход
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'